requirement.txt
pip>=24.0
setuptools>=70.0
wheel
flask
gunicorn
pandas==2.2.2
numpy==1.26.4
pandasai
pandasai-openai
plotly
werkzeug
openpyxl
xlrd
pyarrow
fastparquet
pyyaml
fpdf
reportlab